BENCH: Pin dtype to float64 for np.pad's benchmarks

+++ benchmarks/benchmarks/bench_lib.py
@@ -16,8 +16,8 @@ class Pad(Benchmark):
         [(1000,),
          (10, 100),
          (10, 10, 10),
-         # 50 * 512 * 512 = 13 million points = 46 MB. should be a good
-         # out of cache describing a typical usecase
+         # 50 * 512 * 512 * 64 bits = 100 MiB, corresponds to typical use cases
+         # for large arrays which are out of cache
          (50, 512, 512)],
         [1,
          3,
@@ -34,7 +34,7 @@ class Pad(Benchmark):
     def setup(self, shape, pad_width, mode):
         # Make sure to fill the array to make the OS page fault
         # in the setup phase and not the timed phase
-        self.array = np.full(shape, fill_value=1)
+        self.array = np.full(shape, fill_value=1, dtype=np.float64)
 
     def time_pad(self, shape, pad_width, mode):
         np.pad(self.array, pad_width, mode)
